requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
pydantic==2.10.6
python-dotenv==1.0.1
requests==2.32.3
psycopg2-binary==2.9.9
